pyproject.toml
[project]
name = "llm-for-python-developers"
version = "0.1.0"
description = ""
authors = [
    {name = "tataCrayon",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (==0.115.9)",
    "uvicorn (==0.34.3)",
    "certifi (==2025.1.31)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi-cli (==0.0.8)",
    "fastapi-cloud-cli (==0.1.2)",
    "httpcore (==1.0.8)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "orjson (==3.10.16)",
    "pydantic-extra-types (==2.10.5)",
    "pydantic-settings (==2.9.1)",
    "pygments (==2.19.1)",
    "python-dotenv (==1.1.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "rich (==14.0.0)",
    "rich-toolkit (==0.14.8)",
    "rignore (==0.5.1)",
    "sentry-sdk (==2.32.0)",
    "shellingham (==1.5.4)",
    "truststore (==0.10.0)",
    "typer (==0.16.0)",
    "ujson (==5.10.0)",
    "urllib3 (==2.4.0)",
    "watchfiles (==1.0.5)",
    "websockets (==15.0.1)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.11.18)",
    "aiosignal (==1.3.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "asgiref (==3.8.1)",
    "attrs (==25.3.0)",
    "backoff (==2.2.1)",
    "bcrypt (==4.3.0)",
    "beautifulsoup4 (==4.13.4)",
    "build (==1.2.2.post1)",
    "cachetools (==5.5.2)",
    "charset-normalizer (==3.4.1)",
    "chromadb (==1.0.12)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "coloredlogs (==15.0.1)",
    "dataclasses-json (==0.6.7)",
    "deepseek (==1.0.0)",
    "deprecated (==1.2.18)",
    "distro (==1.9.0)",
    "durationpy (==0.10)",
    "filelock (==3.18.0)",
    "flatbuffers (==25.2.10)",
    "frozenlist (==1.6.0)",
    "fsspec (==2025.5.1)",
    "google-auth (==2.40.2)",
    "google-search-results (==2.4.2)",
    "googleapis-common-protos (==1.70.0)",
    "greenlet (==3.2.0)",
    "grpcio (==1.71.0)",
    "h11 (==0.14.0)",
    "httpx-sse (==0.4.0)",
    "huggingface-hub[hf-xet] (>=0.33.4,<0.34.0)",
    "humanfriendly (==10.0)",
    "idna (==3.10)",
    "importlib-metadata (==8.6.1)",
    "importlib-resources (==6.5.2)",
    "jiter (==0.9.0)",
    "joblib (==1.5.1)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "jsonschema (==4.24.0)",
    "jsonschema-specifications (==2025.4.1)",
    "kubernetes (==32.0.1)",
    "langchain (==0.3.25)",
    "langchain-chroma (==0.2.4)",
    "langchain-community (==0.3.22)",
    "langchain-core (==0.3.62)",
    "langchain-deepseek-official (==0.1.0.post1)",
    "langchain-huggingface (==0.2.0)",
    "langchain-openai (==0.3.14)",
    "langchain-text-splitters (==0.3.8)",
    "langgraph (==0.3.31)",
    "langgraph-checkpoint (==2.0.24)",
    "langgraph-prebuilt (==0.1.8)",
    "langgraph-sdk (==0.1.63)",
    "langsmith (==0.3.33)",
    "marshmallow (==3.26.1)",
    "mmh3 (==5.1.0)",
    "mpmath (==1.3.0)",
    "multidict (==6.4.3)",
    "mypy-extensions (==1.1.0)",
    "networkx (==3.5)",
    "nltk (==3.9.1)",
    "numpy (==2.2.5)",
    "oauthlib (==3.2.2)",
    "onnxruntime (==1.22.0)",
    "openai (==1.75.0)",
    "opentelemetry-api (==1.33.1)",
    "opentelemetry-exporter-otlp-proto-common (==1.33.1)",
    "opentelemetry-exporter-otlp-proto-grpc (==1.33.1)",
    "opentelemetry-instrumentation (==0.54b1)",
    "opentelemetry-instrumentation-asgi (==0.54b1)",
    "opentelemetry-instrumentation-fastapi (==0.54b1)",
    "opentelemetry-proto (==1.33.1)",
    "opentelemetry-sdk (==1.33.1)",
    "opentelemetry-semantic-conventions (==0.54b1)",
    "opentelemetry-util-http (==0.54b1)",
    "ormsgpack (==1.9.1)",
    "overrides (==7.7.0)",
    "packaging (==24.2)",
    "pandas (==2.3.0)",
    "pillow (==11.2.1)",
    "posthog (==4.2.0)",
    "propcache (==0.3.1)",
    "protobuf (==5.29.5)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pydantic (==2.11.3)",
    "pydantic-core (==2.33.1)",
    "pypdf (==5.5.0)",
    "pypika (==0.48.9)",
    "pyproject-hooks (==1.2.0)",
    "pyreadline3 (==3.5.4)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2025.2)",
    "referencing (==0.36.2)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "rpds-py (==0.25.1)",
    "rsa (==4.9.1)",
    "safetensors (==0.5.3)",
    "scikit-learn (==1.6.1)",
    "scipy (==1.15.3)",
    "sentence-transformers (==4.1.0)",
    "setuptools (==80.9.0)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "soupsieve (==2.7)",
    "sqlalchemy (==2.0.40)",
    "starlette (==0.45.3)",
    "sympy (==1.14.0)",
    "tenacity (==9.1.2)",
    "threadpoolctl (==3.6.0)",
    "tiktoken (==0.9.0)",
    "tokenizers (==0.21.1)",
    "torch @ file:///F:/ProgrammingEnvironment/AI/torch-2.7.0%2Bcu128-cp313-cp313-win_amd64.whl",
    "tqdm (==4.67.1)",
    "transformers (==4.52.4)",
    "typing-inspect (==0.9.0)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.2)",
    "tzdata (==2025.2)",
    "websocket-client (==1.8.0)",
    "wrapt (==1.17.2)",
    "xxhash (==3.5.0)",
    "yarl (==1.20.0)",
    "zipp (==3.22.0)",
    "zstandard (==0.23.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "yake (>=0.6.0,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
